+++ src/client/modele/structures/DetailVaisseau.hpp
@@ -13,7 +13,7 @@ struct DetailVaisseau {
     sf::Int32 vieMax; ///< La vie maximum
     sf::Int32 bouclier; ///< Le bouclier courant
     sf::Int32 bouclierMax; ///< Le bouclier maximum
-    float bouclierTaux; ///< Le taux de dégats pris par le bouclier
+    double bouclierTaux; ///< Le taux de dégats pris par le bouclier
     sf::Int32 visibilite; ///< Le rayon de visibilité
     sf::Int32 rayonAttaque; ///< Le rayon d'attaque
     sf::Int32 attaque; ///< L'attaque du vaisseau
