+++ src/old_client/Joueur.cpp
@@ -1,6 +1,6 @@
 #include "Joueur.hpp"
 
-Joueur::Joueur(sf::Int16 _id, std::string _nom, sf::Int32 _energie, sf::Int32 _materiaux) :
+Joueur::Joueur(sf::Int32 _id, std::string _nom, sf::Int32 _energie, sf::Int32 _materiaux) :
     id(_id), nom(_nom), energie(_energie), materiaux(_materiaux) {
 }
 
@@ -8,7 +8,7 @@ const std::string& Joueur::getNom() const {
     return nom;
 }
 
-const sf::Int16& Joueur::getId() const {
+const sf::Int32& Joueur::getId() const {
     return id;
 }
