+++ src/old_client/Joueur.hpp
@@ -5,12 +5,12 @@
 
 class Joueur {
 public:
-    Joueur(sf::Int16 _id = -1, std::string _nom = "Spectateur", sf::Int32 _energie = 0, sf::Int32 _materiaux = 0);
+    Joueur(sf::Int32 _id = -1, std::string _nom = "Spectateur", sf::Int32 _energie = 0, sf::Int32 _materiaux = 0);
     const std::string& getNom() const;
-    const sf::Int16& getId() const;
+    const sf::Int32& getId() const;
     
 private:
-    sf::Int16 id;
+    sf::Int32 id;
     std::string nom;
     sf::Int32 energie;
     sf::Int32 materiaux;
