+++ src/serveur/plateau/NoeudServeur.hpp
@@ -6,17 +6,17 @@
 class NoeudServeur {
 	public:
 		NoeudServeur(Position _position);
-		NoeudServeur(Position& _position, sf::Int16 g, Position _parentPosition);
+		NoeudServeur(Position& _position, sf::Int32 g, Position _parentPosition);
 		bool gInferieur(const NoeudServeur& noeud);
 		bool operator==(const NoeudServeur& noeud);
 		Position getPosition();
 		Position getPositionParent();
-		sf::Int16 getG();
+		sf::Int32 getG();
 	
 	private:
 		Position position;
 		Position positionParent;
-		sf::Int16 distanceG;
+		sf::Int32 distanceG;
 };
 
 #endif // NOEUDSERVEUR_HPP
