+++ src/serveur/structures/batiments/BatimentServeur.hpp
@@ -16,17 +16,17 @@ class BatimentServeur : public Structure {
 public:
     BatimentServeur();
     BatimentServeur(
-        sf::Int16 vieMax, 
-        sf::Int16 bouclierMax, 
-        float bouclierTaux, 
-        sf::Int16 visibilite, 
-        sf::Int16 attaque, 
-        sf::Int16 niveau, 
+        sf::Int32 vieMax, 
+        sf::Int32 bouclierMax, 
+        double bouclierTaux, 
+        sf::Int32 visibilite, 
+        sf::Int32 attaque, 
+        sf::Int32 niveau, 
         TypeBatiment _type = TypeBatiment::Inexistant
     );
     TypeBatiment getType() const;
-    sf::Int16 getNiveau();
-    void setNiveau(sf::Int16 niveau);
+    sf::Int32 getNiveau();
+    void setNiveau(sf::Int32 niveau);
     void subir(VaisseauServeur const& attaquant);
     void subir(BatimentServeur const& attaquant);
 
@@ -34,7 +34,7 @@ protected:
 
 private:
     TypeBatiment type;
-    sf::Int16 niveau;
+    sf::Int32 niveau;
 
     friend sf::Packet& operator <<(sf::Packet& paquet, const BatimentServeur& batiment);
 };
