+++ src/serveur/structures/vaisseaux/VaisseauServeur.hpp
@@ -21,25 +21,25 @@ class VaisseauServeur : public Structure {
 public:
     VaisseauServeur();
     VaisseauServeur(
-        sf::Int16 _vieMax,
-        sf::Int16 _bouclierMax,
-        float _bouclierTaux,
-        sf::Int16 _visibilite,
-        sf::Int16 _attaque,
-        sf::Int16 _distanceMax,
-        sf::Int16 _consommation,
+        sf::Int32 _vieMax,
+        sf::Int32 _bouclierMax,
+        double _bouclierTaux,
+        sf::Int32 _visibilite,
+        sf::Int32 _attaque,
+        sf::Int32 _distanceMax,
+        sf::Int32 _consommation,
         TypeVaisseau _type = TypeVaisseau::Inexistant
     );
 
     // Getters
-    sf::Int16 getConsommation() const;
-    sf::Int16 getDistanceMax() const;
+    sf::Int32 getConsommation() const;
+    sf::Int32 getDistanceMax() const;
     TypeVaisseau getType() const;
 
     // Setters
-    void setConsommation(const sf::Int16 _consommation);
+    void setConsommation(const sf::Int32 _consommation);
 
-    void modifierEnergie(sf::Int16 const valeur);
+    void modifierEnergie(sf::Int32 const valeur);
     static VaisseauServeur cloner(VaisseauServeur const& modele, TechnologieStructure techS, TechnologieVaisseau techV);
     int triangulaire(VaisseauServeurPtr vAttaquant);
     void afficher(std::ostream& fluxSortant) const;
@@ -48,8 +48,8 @@ protected:
 
 private:
     TypeVaisseau type;
-    sf::Int16 distanceMax;
-    sf::Int16 consommation;
+    sf::Int32 distanceMax;
+    sf::Int32 consommation;
     
     friend std::ostream& operator<<(std::ostream& fluxSortant, VaisseauServeur const& VaisseauServeur);
     friend sf::Packet& operator <<(sf::Packet& paquet, const VaisseauServeur& vaisseau);
